requirements.txt
streamlit
streamlit-webrtc
opencv-python-headless
deepface
tf-keras
langchain
langchain-groq
langgraph
langchain-community
sqlalchemy
pandas  
plotly
python-dotenv
fpdf2
schedule
Pillow
numpy
groq